v1/requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
drf-spectacular>=0.27
django-environ>=0.11
celery[redis]>=5.3
redis>=5.0
elasticsearch>=8.13,<9.0
selenium>=4.18
webdriver-manager>=4.0
requests>=2.31
zstandard>=0.22
brotli>=1.1
urllib3>=2.0
watchdog>=4.0
